refactor: move Interface into models repo
